+++ x-pack/test/security_solution_cypress/cypress/cypress_ci.config.ts
@@ -33,7 +33,6 @@ export default defineCypressConfig({
   viewportHeight: 946,
   viewportWidth: 1680,
   e2e: {
-    supportFile: 'cypress/support/ess_e2e.ts',
     baseUrl: 'http://localhost:5601',
     experimentalMemoryManagement: true,
     experimentalCspAllowList: ['default-src', 'script-src', 'script-src-elem'],
