--- x-pack/test/security_solution_cypress/cypress/support/serverless_e2e.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import './commands';
-import 'cypress-real-events/support';
-import registerCypressGrep from '@cypress/grep';
-
-before(() => {
-  cy.task('esArchiverLoad', { archiveName: 'auditbeat' });
-});
-
-registerCypressGrep();
-
-Cypress.on('uncaught:exception', () => {
-  return false;
-});
